src/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery-reservation>=1.1.0
google-api-core>=1.26.2
